// src/types/values.ts
export type values = {
  Point: number,                          // 1.ポイント
  CriticalHit: number,                    // 2.クリティカル
  AllDamage: number,                      // 3.すべてのダメージ
  Accuracy: number,                       // 4.命中
  DoubleChance: number,                   // 5.ダブル確率
  ExtraPvEDamage: number,                 // 6.PvE追加ダメージ
  IgnoreDamageResistance: number,         // 7.ダメージ耐性無視
  ExtraPvPDamage: number,                 // 8.PvP追加ダメージ
  IgnoreExtraReductionToStunned: number,  // 9.スタン対象追加リダクション無視
  MaxHP: number,                          // 10.最大HP
  PvPWeaponDefense: number,               // 11.PvP武器ダメージ耐性
  CriticalHitResistance: number,          // 12.クリティカル抵抗
  PvPSkillDefense: number,                // 13.PvPスキルダメージ耐性
  WeaponBlock: number,                    // 14.武器防御
  DamageReduction: number,                // 15.ダメージリダクション
  Defense: number,                        // 16.防御力
  ExtraDamageOnCriticalHit: number,       // 17.クリティカル時の追加ダメージ
  SkillDamageBoost: number,               // 18.スキルダメージ増幅
  WeaponDamageBoost: number,              // 19.武器ダメージ増幅
  AllTypeBoost: number,                   // 20.すべての属性増幅
  BlockPenetration: number,               // 21.防御貫通
  PvPDamageResistance: number,            // 22.PvPダメージ耐性
  PotionRecoveryRate: number,             // 23.ポーション回復率
  HealBoost: number,                      // 24.ヒール増幅
  DoubleResistance: number,               // 25.ダブル抵抗
  DamageResistance: number,               // 26.ダメージ耐性
  CCAccuracy: number,                     // 27.状態異常命中
  PvEAccuracy: number,                    // 28.PvE命中
  ExtraAccuracyToStunned: number,         // 29.スタン対象追加命中
  PvPAccuracy: number,                    // 30.PvP命中
  SkillAccuracy: number,                  // 31.スキル命中
  AbsoluteAccuracy: number,               // 32.絶対命中率
  StunAccuracy: number,                   // 33.スタン命中
  HoldAccuracy: number,                   // 34.ホールド命中
  SkillResistance: number,                // 35.スキル抵抗
  CCDurationReduction: number,            // 36.状態異常時間減少
  HoldResistance: number,                 // 37.ホールド耐性
  WeaponDefense: number,                  // 38.武器ダメージ耐性
  SkillDefense: number,                   // 39.スキルダメージ耐性
  StunResistance: number,                 // 40.スタン耐性
  AggressionResistance: number,           // 41.ヘイト耐性
  SilenceResistance: number,              // 42.沈黙耐性
  CCStackResistance: number,              // 43.重複状態異常耐性
  BlessingsRechargeIncrease: number,        // 44.祝福増加量増加
  BasicAttackDamage: number,                // 45.通常攻撃ダメージ
  BasicAttackCriticalHit: number,           // 46.通常攻撃クリティカル
  BasicAttackDamageResistance: number,      // 47.通常攻撃ダメージ耐性
  BasicAttackDamageReduction: number,       // 48.通常攻撃ダメージリダクション
  BasicAttackCriticalHitResistance: number, // 49.通常攻撃クリティカル抵抗
  BlessingsConservation: number,            // 50.祝福消耗量減少
  BasicAttackDamageBoost: number,           // 51.通常攻撃ダメージ増幅
  BonusEXP: number,                         // 52.EXPボーナス
}[];